requirements.txt
fastapi
uvicorn[standard]
torch
transformers
diffusers
accelerate
safetensors
pillow
rembg
realesrgan
gfpgan
controlnet_aux
python-dotenv
